refactor(utils): move hash & rand into utils/crypto
