+++ client/highway.go
@@ -12,8 +12,8 @@ import (
 
 	highway2 "github.com/LagrangeDev/LagrangeGo/packets/highway"
 	"github.com/LagrangeDev/LagrangeGo/packets/pb/service/highway"
-	"github.com/LagrangeDev/LagrangeGo/utils"
 	"github.com/LagrangeDev/LagrangeGo/utils/binary"
+	"github.com/LagrangeDev/LagrangeGo/utils/crypto"
 	"github.com/RomiChan/protobuf/proto"
 )
 
@@ -137,7 +137,7 @@ func (c *QQClient) SendUpBlockAsync(block UpBlock, server string) bool {
 		DataFlag:   16,
 		CommandId:  uint32(block.CommandId),
 	}
-	md5 := utils.MD5Digest(block.Block)
+	md5 := crypto.MD5Digest(block.Block)
 	segHead := &highway.SegHead{
 		ServiceId:     proto.Some(uint32(0)),
 		Filesize:      block.FileSize,
