+++ info/device.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"strconv"
 
+	"github.com/LagrangeDev/LagrangeGo/utils/crypto"
 	"github.com/LagrangeDev/LagrangeGo/utils/platform"
 
 	"github.com/LagrangeDev/LagrangeGo/utils"
@@ -18,8 +19,8 @@ type DeviceInfo struct {
 
 func NewDeviceInfo(uin int) *DeviceInfo {
 	return &DeviceInfo{
-		Guid:          fmt.Sprintf("%X", utils.MD5Digest(utils.S2B(strconv.Itoa(uin)))),
-		DeviceName:    fmt.Sprintf("Lagrange-%X", utils.MD5Digest(utils.S2B(strconv.Itoa(uin)))[0:4]),
+		Guid:          fmt.Sprintf("%X", crypto.MD5Digest(utils.S2B(strconv.Itoa(uin)))),
+		DeviceName:    fmt.Sprintf("Lagrange-%X", crypto.MD5Digest(utils.S2B(strconv.Itoa(uin)))[0:4]),
 		SystemKernel:  fmt.Sprintf("%s %s", platform.GetSystem(), platform.GetVersion()),
 		KernelVersion: platform.GetVersion(),
 	}
